fastapi/requirements.txt
fastapi[standard]>=0.113.0,<0.114.0
uvicorn[standard]>=0.27.0,<0.28.0
SQLAlchemy>=2.0.0
psycopg2-binary
python-dotenv
